Preserve the audio player state
First, ensure that the `[id="player"]` element is rendered with
[data-turbo-permanent][] so that the same element instance is preserved
as Turbo Drive navigates the rest of the application.

Next, change the `[id="audio"]` element nested within the
`[id="player"]` from a `<div>` into a [`<turbo-frame>`
element][turbo-frame]. We'll drive that `<turbo-frame>` with some of our
application's `<form>` elements, namely the ones that submit requests to
the `GET /podcasts/:podcast_id/episodes/:id` route.

To do so, we'll render those `<form>` elements with the
[data-turbo-frame][] attribute set to `"audio"`. To prevent multiple
submissions from re-loading the frame, we'll also introduce the
`application#preventDefault` action with a [custom `:reload` Action
Option][custom-action-option]. To grant that controller access to the
entire document, we'll render it on the page's `<html>` element.

The `preventDefault` action cancels the `submit` event. When combined
with the `:reload` Action Option, it will only ever be invoked when the
[SubmitEvent][] that's dispatched indicates that the `<form>` submission
will drive the document or a targeted `<turbo-frame>` to an
already-loaded URL. If that's the case, the Action Option will prevent
the `submit` event from being dispatched to `application#preventDefault`.

Finally, to control an already-loaded `<turbo-frame id="audio">` element
that nests an `<audio>` element, this commit also introduces a `player`
controller to manage the `<audio>` element's state, and a `play-button`
Stimulus controller to drive those state changes.

The `play-button` controller utilizes the concept of an [Outlet][] to
maintain direct access to a `player` Stimulus controller. The Outlet
resolves the correct `<audio>` element based on a CSS selector that
targets it by an `[id]` attribute generated by a call to
`dom_id($EPISODE_INSTANCE, :audio)`. The `play-button` controller
synchronizes the `<button>` element's `[aria-controls]` and
`[aria-pressed]` attribute state with any `player`-side state changes,
and routes `click` events to its own `play-button#toggle` action.

[data-turbo-permanent]: https://turbo.hotwired.dev/handbook/building#persisting-elements-across-page-loads
[turbo-frame]: https://turbo.hotwired.dev/handbook/frames
[data-turbo-frame]: https://turbo.hotwired.dev/handbook/frames#targeting-navigation-into-or-out-of-a-frame
[custom-action-option]: https://stimulus.hotwired.dev/reference/actions#options
[SubmitEvent]: https://developer.mozilla.org/en-US/docs/Web/API/SubmitEvent
[Outlet]: https://stimulus.hotwired.dev/reference/outlets