Implement pages with Tailwind UI Template
First, depend on the [inline_svg][] gem to support rendering `<svg>`
elements based on files in the `app/views/images` directory.

Next, introduce the `EpisodesController#index` and
`EpisodesController#show`, along with the corresponding routes and
views, to render the available `Episode` records. Also declare a
`PodcastsController#index` action to find the first available
`Podcast` and redirect to its list of `Episode` records.

To style the pages, borrow from the [Transmit][] Tailwind template theme
(along with its [preview][]). Replace all its static information with
dynamic values read from `Podcast` and `Episode` records.

This commit also introduces a suite of integration tests to guarantee
that changes to the page's structure preserve its content.

Additionally, we use a [tag helper][] to build the play button in an
effort to emphasize the `diff` in a later lesson.

[inline_svg]: https://tailwindui.com/templates/transmit
[Transmit]: https://tailwindui.com/templates/transmit
[preview]: https://transmit.tailwindui.com
[tag helper]: https://api.rubyonrails.org/classes/ActionView/Helpers/TagHelper.html