Episode Search
First, we'll implement the `Episode.containing` scope to transform a
search term into an SQL [ILIKE][] query.

Next, create the `Search` class to serve as a model for a term-based
`Episode` query. We'll construct instances based on the `query` and
`page` parameters submitted as `URLSearchParams`.

Then, we'll declare the `Search#episodes` method to construct an
appropriate `Episode.containing` query, along with a
`Search#search_results` method to paginate the collection of `Episode`
records and transform them into `SearchResult` instances. Both `Search`
and `SearchResult` inherit from the new `ApplicationModel` (not to be
confused with `ApplicationRecord`) base class.

Finally, we'll introduce the `/podcasts/:podcast_id/search_results`
route along with the `SearchResultsController#index` action to fetch
results and render them as a list of `search_results/search_result`
partials. The `search_results/search_result` partial composes calls to
the [highlight][] and [excerpt][] view helpers to find and highlight
portions of the result that match the search term.

For now, we'll render a [mirage][] search field in the main navigation
to link to the `search_results#index` route and template. In the future,
we'll be able to progressively enhance search to be globally accessible
throughout the application.

[ILIKE]: https://www.postgresql.org/docs/current/functions-matching.html#FUNCTIONS-LIKE
[highlight]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/TextHelper.html#method-i-highlight
[excerpt]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/TextHelper.html#method-i-excerpt
[mirage]: https://en.wikipedia.org/wiki/Mirage

Co-authored-by: Steve Polito <stevepolito@hey.com>
Co-authored-by: Sean Doyle <sean.p.doyle24@gmail.com>