[BACKPORT]: Promote lazily-loaded Turbo Frame navigation
The changed merged into `@hotwired/turbo` as part of
[hotwired/turbo#790][] resolve an important issue around how
`<turbo-frame loading="lazy">` elements treat `[data-turbo-action]`
attributes.

With #790 merged into `main`, it is not yet part of an official
`@hotwired/turbo` or `@hotwired/turbo-rails` release, and is only
available through a manual backport.

Once `@hotwired/turbo@7.2.5` is released, this commit can be reverted or
deleted from the history.

[hotwired/turbo#790]: https://github.com/hotwired/turbo/pull/790