Lesson 3: Infinite scroll
This commit progressively enhances the episodes index by automatically
scrolling to the next page of episodes by leveraging [Turbo Frames][]
and [Stimulus][].

We set `[data-turbo-action]` to `replace` in order to [_replace_][]
visit history while scrolling. In other words, if the user were to click
the back button, the browser would not navigate to the previous page,
but instead would navigate to the previously visited page.

[Turbo Frames]: https://turbo.hotwired.dev/handbook/frames
[Stimulus]: https://stimulus.hotwired.dev
[_replace_]: https://turbo.hotwired.dev/handbook/drive#application-visits

Co-authored-by: Sean Doyle <sean.p.doyle24@gmail.com>