Lesson 4: Typeahead Search
This commit extends the `element` Stimulus controller with a
`requestSubmit` action. We'll render the `<form>` element with
`[data-controller="element"]` and
`[data-action="debounced:input->element#requestSubmit"]` to route every
[input][] event to the `element#requestSubmit` action, which will submit
the `<form>`.

The `debounced:input` portion of the `[data-action]` attribute is refers
to an event that's dispatched by the [debounced][] package, and is named
after the built-in [input][] event.

To preserve the query and its focus state throughout the live-search,
we'll render the `<input>` element with the [data-turbo-permanent][]
attribute. When combined with an `[id]` attribute that's consistent
across the requesting document and the response body, Turbo Drive will
carry the element instance forward through the navigation, and backward
through a history restoration.

We'll provide a `debounced` configuration to the page by rendering an
in-line `<script type="module">` element that invokes
`debounced.initialize` with a `JSON` object generated from values read
from a call to [`config_for(:debounced)`][config_for]. In most
environments, the `<form>` will wait 100 milliseconds between `input`
events. In `test`, it'll submit immediately.

Since the `<form>` will be making rapid submissions and those
submissions will result in a sequence of Turbo Drive visits, we'll want
to render the `<form>` with [data-turbo-action="replace"][] so that we
_replace_ the current History entry, instead of creating a new entry for
each and every keystroke.

[input]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/input_event
[data-turbo-permanent]: https://turbo.hotwired.dev/handbook/building#persisting-elements-across-page-loads
[debounced]: https://github.com/hopsoft/debounced#why
[config_for]: https://edgeapi.rubyonrails.org/classes/Rails/Application.html#method-i-config_for
[data-turbo-action="replace"]: https://turbo.hotwired.dev/handbook/drive#application-visits