Keyboard Shortcuts: Our first lines of JavaScript
We'll ease into the concept how Hotwire by progressively enhancing the
application's Search page navigation with a <kbd><kbd>⌘</kbd>
<kbd>k</kbd></kbd>-powered keyboard shortcut.

This commit re-introduces the `layouts/application.html.erb` template's
call to [javascript_importmap_tags][] to load our application's
[importmap]-backed JavaScript bundle.

We'll attach a new `hotkey` controller to the document's `<body>`
element so that we can integrate with the [@github/hotkey][] package.
Next, we'll mark the Search navigation with the
`[data-hotkey-target="shortcut"]` attribute, along with the
`[data-hotkey="Meta+k"]` to attach a keyboard listener.

In the style of its Rails-side counterparts, the `hotkey` controller
imports and extends a base level `ApplicationController`. The
`ApplicationController` will start as an empty class, but provides an
opportunity to define shared actions, logic, or utilities.

[javascript_importmap_tags]: https://github.com/rails/importmap-rails#preloading-pinned-modules
[importmap]: https://github.com/WICG/import-maps
[@github/hotkey]: https://github.com/WICG/import-maps