Build assets during setup (#16)
There appears to be a [bug][1] with tailwindcss-rails, or possibly
Rails, that is affecting how assets are precompiled on a cloned
repository.

Prior to this commit, we would get the following error after cloning the
repo and running the test suite:

```text
ActionView::Template::Error: The asset 'tailwind.css' was not found in
the load path.
```
One solution is to run `test:prepare` in the setup script which will
generate the assets. There is an [open pull request][2] against Rails
that would introduce this change based on [similar feedback][3]

[1]: https://github.com/rails/tailwindcss-rails/issues/158
[2]: https://github.com/rails/rails/pull/47719
[3]: https://github.com/rails/rails/pull/47210#issuecomment-1476420481