Mount good_jobs UI
When initiating the app for the first time,
and accessing localhost:3000, it can be confusing
to get errors about episodes not being found, or
their images not being found.

This happens because in the setup, background jobs
to download the audio and images are scheduled.
When we run `./bin/dev`, the background jobs
will get executed. This will take some time, so
I thought of adding a note to users.