Update README
Updates the intention of this project by removing language around
testing. For now, we felt that was a blocker to releasing this resource.

Adds additional resources located on our marketing site.

Also adds a footer based on our [templates][]. Because [contributing][]
to this project requires keeping a specific git history, we opted to not
[dynamically][] generate the footer.

[templates]: https://github.com/thoughtbot/templates
[contributing]: https://github.com/thoughtbot/botcasts/blob/main/CONTRIBUTING.md
[dynamically]: https://github.com/thoughtbot/templates?tab=readme-ov-file#dynamic-content-rendering-for-the-footer